% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_relative_effort_fleet_stock.R
\name{plot_relEffortFltStk}
\alias{plot_relEffortFltStk}
\title{Relative effort for fleets and stock}
\usage{
plot_relEffortFltStk(
  data,
  limits = c(-100, 100),
  xlab = "Stock",
  ylab = "Fleet",
  fillLegendTitle = "Variation\\n in effort"
)
}
\arguments{
\item{data}{}

\item{limits}{}

\item{xlab}{}

\item{ylab}{}

\item{fillLegendTitle}{}
}
\description{
Relative effort for fleets and stock
}
